refactor(retention): move the decision core to a shared include root (#2549)
audit_retention_rules.hpp (classify/Facts/moved_at_least) is a pure
function with no store type or I/O — the right shape to share between
server/core and agents/, which have no common include root today.
Moves it to common/include/yuzu/ unchanged (name and namespace kept
verbatim to avoid call-site churn) and wires the new include dir into
server/core, agents/core, and agents/plugins/tar so the upcoming
TAR/Guardian-journal latch fixes (#2573) can adopt it instead of
hand-porting a fourth and fifth copy.

Behaviour-preserving: no logic changed, only the header's location.
