+++ server/core/src/guaranteed_state_store.cpp
@@ -1,6 +1,6 @@
 #include "guaranteed_state_store.hpp"
 
-#include "audit_retention_rules.hpp"
+#include <yuzu/audit_retention_rules.hpp>
 #include "pg/pg_exec.hpp"
 #include "pg/pg_migration_runner.hpp"
 #include "pg/pg_pool.hpp"
