+++ tests/unit/server/test_audit_store.cpp
@@ -12,7 +12,7 @@
  */
 
 #include "audit_store.hpp"
-#include "audit_retention_rules.hpp"
+#include <yuzu/audit_retention_rules.hpp>
 
 #include "pg/pg_exec.hpp"
 #include "pg/pg_pool.hpp"
